POLIO-697: move tabs to separate component
